Modifications to my test class
Did a major overhaul of all of my  trap classes and added a bunch of
textures to can draw.  Changes the checkerboard class so it draws
textures. Too see a basic demo please run my test class.  To make the
traps move press the r or f keys respectively.  This action will be pass
to the button in the future but again this is just a demo.